Just pushed some changes for it to use jwt public and secrets in its own file.
