+++ utilities/jwt.js
@@ -1,0 +1,36 @@
+const jwt = require('jsonwebtoken');
+const {jwtSecret, jwtPublic} = require('./secrets');
+
+function generateToken(user){
+    const payload = {
+        subject: user.id,
+        email: user.email
+    };
+
+    const options = {
+        issuer: 'Workout Tracker',
+        subject: user.id,
+        audience: 'WorkoutTracker.com',
+        expiresIn: '24h',
+        algorithm: 'RS256'
+    };
+
+    return jwt.sign(payload, jwtSecret, options);
+}
+
+function verifyToken(user){
+    const verifyOptions = {
+        issuer: 'Workout Tracker',
+        subject: user.id,
+        audience: 'WorkoutTracker.com',
+        maxAge: '24h',
+        algorithm: ["RS256"]
+    }
+
+    return jwt.verify(token, jwtPublic, verifyOptions);
+}
+
+module.exports = {
+    generateToken,
+    verifyToken
+}
